// src/app/models/event.model.ts
/*
import { Pass } from "./pass.model";
import { Feedback } from "./feedback.model";

export interface Event {
  id: number;
  title: string;
  description: string;
  venue: string;
  date: string;
  duration: number;

 
  
  }
  export enum EventPass {
    TYPE1 = 'STANDARD',
    TYPE2 = 'VIP',
    TYPE3 = 'PREMIUM',
  }
  */